requirements.txt
PySide6
qtawesome
psutil
SQLAlchemy
requests
WMI
pythonnet
groq
tenacity
pywin32
PyYAML
